requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.0
google-generativeai==0.8.3
